Refactor call state servers and unify Twilio/Telnyx logic
Extract shared functionality into a `TwilioBaseCallStateServer` and modularize Telnyx and Twilio-specific components. Introduce `RequestVerificationMiddlewareService` for pluggable request verification and restructure dependencies for better separation of concerns.
