+++ core/src/scala/io/github/nafg/dialoguestate/CallState.scala
@@ -1,9 +1,8 @@
 package io.github.nafg.dialoguestate
 
 sealed trait CallState
-
 object CallState {
-  case class Const(callTree: CallTree)                                                                 extends CallState
-  case class Digits(gather: CallTree.Gather, handleDigits: String => CallTree.Callback)                extends CallState
-  case class Recording(record: CallTree.Record, handleRecording: RecordingResult => CallTree.Callback) extends CallState
+  case class Ready(callTree: CallTree)                  extends CallState
+  case class AwaitingDigits(gather: CallTree.Gather)    extends CallState
+  case class AwaitingRecording(record: CallTree.Record) extends CallState
 }
