+++ core/src/scala/io/github/nafg/dialoguestate/RequestVerificationMiddlewareService.scala
@@ -1,0 +1,12 @@
+package io.github.nafg.dialoguestate
+
+import zio.http.Middleware
+
+trait RequestVerificationMiddlewareService  {
+  def middleware: Middleware[Any]
+}
+object RequestVerificationMiddlewareService {
+  object NoVerification extends RequestVerificationMiddlewareService {
+    override def middleware: Middleware[Any] = Middleware.identity
+  }
+}
