+++ twilio-base/src/scala/io/github/nafg/dialoguestate/twilio/base/Voice.scala
@@ -1,0 +1,3 @@
+package io.github.nafg.dialoguestate.twilio.base
+
+case class Voice(value: String)
