--- twilio/src/scala/io/github/nafg/dialoguestate/twilio/Voice.scala
@@ -1,3 +1,0 @@
-package io.github.nafg.dialoguestate.twilio
-
-case class Voice(value: String)
